Extract producer retry test to separate spec
