+++ jms/src/test/scala/akka/stream/alpakka/jms/JmsSpec.scala
@@ -7,9 +7,13 @@ package akka.stream.alpakka.jms
 import akka.actor.ActorSystem
 import akka.stream.{ActorMaterializer, ActorMaterializerSettings, Supervision}
 import akka.testkit.{SocketUtil, TestKit}
+import javax.jms._
 import org.apache.activemq.broker.BrokerService
+import org.mockito.ArgumentMatchers.{any, anyBoolean, anyInt}
+import org.mockito.Mockito.when
 import org.scalatest._
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
+import org.scalatest.mockito.MockitoSugar
 
 abstract class JmsSpec
     extends WordSpec
@@ -17,7 +21,8 @@ abstract class JmsSpec
     with BeforeAndAfterAll
     with BeforeAndAfterEach
     with ScalaFutures
-    with Eventually {
+    with Eventually
+    with MockitoSugar {
 
   implicit val system = ActorSystem(this.getClass.getSimpleName)
 
@@ -57,4 +62,14 @@ abstract class JmsSpec
 
   case class Context(url: String, broker: BrokerService)
 
+  def withMockedProducer(test: ProducerMock => Unit): Unit = test(ProducerMock())
+
+  case class ProducerMock(factory: ConnectionFactory = mock[ConnectionFactory],
+                          connection: Connection = mock[Connection],
+                          session: Session = mock[Session],
+                          producer: MessageProducer = mock[MessageProducer]) {
+    when(factory.createConnection()).thenReturn(connection)
+    when(connection.createSession(anyBoolean(), anyInt())).thenReturn(session)
+    when(session.createProducer(any[javax.jms.Destination])).thenReturn(producer)
+  }
 }
